rag-with-azuresearch/requirements.txt
langchain
langchain-community
langchain-openai
azure-search-documents
azure-identity
python-dotenv
streamlit
unstructured
psutil